requirements.txt
# RAGシステム用の依存パッケージ
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
pydantic>=2.4.2

langchain-community==0.3.19
langchain-core==0.3.41
langchain-cohere==0.4.2
lxml==5.3.1
html2text==2020.1.16
tiktoken==0.9.0 